02_modesetting: fix byte order in ARGB example
This is about endianness, so this going to be confusing.

We want to express the colour #112233, ie. R=0x11, G=0x22, B=0x33,
A=0xFF.

The DRM format we use is ARGB8888, which means: A at the highest
address, then R, G, then B at the lowest address. This gives us:

    uint32_t argb = 0xFF112233;
    //                ^ highest address
    //       lowest address ^

Then we need to convert to the memory layout, little endian inverts the
order:

    uint8_t pixel[4] = { 0x33, 0x22, 0x11, 0xFF };
    //                   ^ lowest address
    //                     highest address ^

See also: https://afrantzis.com/pixel-format-guide/drm.html